Split out whitespace rules into multiple rules. Split the overflow indentation check into its own rule.
