+++ lib/rules/indentation.js
@@ -196,14 +196,11 @@ module.exports = {
 			node.body.forEach (inspectBlockItem.bind (null, blockIndent));
 		});
 
-
-
 		context.on ('StructDeclaration', function (emitted) {
 			var node = emitted.node,
 				body = node.body || [],
 				endingLineNum = sourceCode.getEndingLine (node);
 			var structDeclarationLineText, currentIndent, currentIndentLevel, structIndent;
-			var MAX_ATTR_IN_SINGLE_LINE = 3;
 
 			function inspectStructAttr (structIndent, attr) {
 				var indentRegExp = new RegExp ('^' + structIndent + '[^\\s(\/\*)]'),
@@ -217,17 +214,6 @@ module.exports = {
 			}
 
 			if (emitted.exit) {
-				return;
-			}
-
-			//raise an error and stop linting if more than 3 attributes exist & declaration is on single line
-			if (sourceCode.getLine (node) === endingLineNum) {
-				if (body.length > MAX_ATTR_IN_SINGLE_LINE) {
-					context.report ({
-						node: node,
-						message: '\'' + node.name + '\': In case of more than 3 properties, struct declaration needs to be spread over multiple lines with 1 property per line.'
-					});
-				}
 				return;
 			}
 
@@ -256,7 +242,6 @@ module.exports = {
 			var node = emitted.node, elements = node.elements;
 			var endingLineNum = sourceCode.getEndingLine (node),
 				arrayExpressionLineText, currentIndent, currentIndentLevel, arrayIndent;
-			var MAX_ELEMS_IN_SINGLE_LINE = 3;
 
 			function inspectElement (arrayIndent, elem) {
 				var indentRegExp = new RegExp ('^' + arrayIndent + '[^\\s(\/\*)]'),
@@ -270,16 +255,6 @@ module.exports = {
 			}
 
 			if (emitted.exit) {
-				return;
-			}
-
-			if (sourceCode.getLine (node) === endingLineNum) {
-				if (elements.length > MAX_ELEMS_IN_SINGLE_LINE) {
-					context.report ({
-						node: node,
-						message: 'In case of more than 3 elements, array expression needs to be spread over multiple lines with 1 element per line.'
-					});
-				}
 				return;
 			}
 
@@ -305,7 +280,6 @@ module.exports = {
 		//function params (if on multiple lines)
 		context.on ('FunctionDeclaration', function (emitted) {
 			var node = emitted.node, params = node.params || [];
-			var MAX_PARAMS_ON_SINGLE_LINE = 3;
 
 			var startLine = sourceCode.getLine (node),
 				lastArgLine = params.length ? sourceCode.getEndingLine (params.slice (-1) [0]) : startLine,
@@ -323,16 +297,6 @@ module.exports = {
 			}
 
 			if (emitted.exit) {
-				return;
-			}
-
-			if (startLine === lastArgLine) {
-				if (params.length > MAX_PARAMS_ON_SINGLE_LINE) {
-					context.report ({
-						node: node,
-						message: 'Function \'' + node.name + '\': in case of more than 3 parameters, drop each into its own line.'
-					});
-				}
 				return;
 			}
 
@@ -361,7 +325,6 @@ module.exports = {
 			var node = emitted.node;
 			var endingLineNum = sourceCode.getEndingLine (node),
 				callExpressionLineText, currentIndent, currentIndentLevel, argIndent;
-			var MAX_ARGS_ON_SINGLE_LINE = 3;
 
 			function inspectArgument (argIndent, argument) {
 				var indentRegExp = new RegExp ('^' + argIndent + '[^\\s(\/\*)]'),
@@ -375,16 +338,6 @@ module.exports = {
 			}
 
 			if (emitted.exit) {
-				return;
-			}
-
-			if (sourceCode.getLine (node) === endingLineNum) {
-				if (node.arguments.length > MAX_ARGS_ON_SINGLE_LINE) {
-					context.report ({
-						node: node,
-						message: 'Function \'' + node.callee.name + '\': in case of more than 3 arguments, drop each into its own line.'
-					});
-				}
 				return;
 			}
