+++ lib/rules/operator-whitespace.js
@@ -11,6 +11,54 @@ module.exports = {
 
 		var sourceCode = context.getSourceCode ();
 
+		context.on ('AssignmentExpression', function (emitted) {
+			/**
+			 * node.operator is refined here by adding backslash before all the 'special' characters.
+			 * 'special' chars are thos chars that are part of solidity assignment operators and, if used without backslash in JS RegExp,
+			 * behave as wildcard characters. So to make sure they're treated as simple strings, we add '\' before them.
+			 * As of today, these chars include: * / + | ^
+			 */
+			var node = emitted.node,
+				op = node.operator.replace (/([\+\*\/\|\^])/g, '\\$1'), opLength = node.operator.length;
+
+			if (emitted.exit) {
+				return;
+			}
+
+			// If expression is 'abc *= def;', then charsAfterLeftNode will contain ' *= d'.
+			var charsAfterLeftNode = sourceCode.getNextChars (node.left, 3 + opLength),
+				validationRegexp = new RegExp ('^ ' + op + ' [^\\s]$');
+
+			(!validationRegexp.test (charsAfterLeftNode)) && context.report ({
+				node: node.left,
+				message: 'Assignment operator must have exactly single space on both sides of it.'
+			});
+		});
+
+		//statement like `var x = 10` doesn't come under AssignmentExpression, so needs to be checked separately
+		context.on ('VariableDeclaration', function (emitted) {
+			var node = emitted.node, code = sourceCode.getText (node);
+
+			if (emitted.exit) {
+				return;
+			}
+
+			//if a particular character is '=', check its left and right for single space
+			for (var i = 2; i < code.length; i++) {
+				if (code [i] === '=') {
+					(!/^[^\/\s] $/.test (code.slice (i-2, i))) && context.report ({
+						node: node,
+						message: 'There should be only a single space between assignment operator \'=\' and its left side.' 
+					});
+
+					(!/^ [^\/\s]$/.test (code.slice (i+1, i+3))) && context.report ({
+						node: node,
+						message: 'There should be only a single space between assignment operator \'=\' and its right side.' 
+					});
+				}
+			}
+		});
+
 		context.on ('BinaryExpression', function (emitted) {
 			var leftNode,
 			    node = emitted.node;
