+++ lib/rules/whitespace.js
@@ -11,78 +11,6 @@ module.exports = {
 
 		var sourceCode = context.getSourceCode ();
 
-		//same could potentially be applied to FunctionDeclaration
-		context.on ('CallExpression', function (emitted) {
-			var node = emitted.node,
-				callArgs = node.arguments;
-
-			function inspectCallArgForCommaWhitespace (arg) {
-				var charAfterArg = sourceCode.getNextChar (arg);
-
-				(charAfterArg !== ',') && context.report ({
-					node: arg,
-					location: {
-						column: sourceCode.getEndingColumn (arg) + 1
-					},
-					message: 'There should be no whitespace or comments between argument and the comma following it.'
-				});
-			}
-
-			if (emitted.exit) {
-				return;
-			}
-
-			var nodeCode = sourceCode.getText (node);
-
-			//for a 0-argument call, ensure that name is followed by '()'
-			if (!callArgs.length) {
-				for (var i = nodeCode.length; i > 0; i--) {
-					if (nodeCode [i] === ')' && nodeCode [i-1] === '(') {
-						return;
-					}
-					if (/[\s\(\)]/.test (nodeCode [i])) {
-						break;
-					}
-				}
-
-				return context.report ({
-					node: node,
-					message: '\"' + nodeCode + '\": ' +
-					'A call without arguments should have brackets without any whitespace between them, like \'functionName ()\'.'
-				});
-			}
-
-			//CHECKING FOR COMMA WHITESPACE
-			callArgs.slice (0, -1).forEach (inspectCallArgForCommaWhitespace);
-
-			var lastCallArg = callArgs.slice (-1) [0];
-
-			//if call spans over multiple lines (due to too many arguments), below rules don't apply
-			if (sourceCode.getLine (node) !== sourceCode.getEndingLine (lastCallArg)) {
-				return;
-			}
-
-			var charBeforeFirstArg = sourceCode.getPrevChar (callArgs [0]),
-				charAfterLastCallArg = sourceCode.getNextChar (lastCallArg);
-
-			(callArgs [0].type !== 'NameValueAssignment' && charBeforeFirstArg !== '(') && context.report ({
-				node: callArgs [0],
-				location: {
-					column: sourceCode.getColumn (callArgs [0]) - 1
-				},
-				message: '\'' + node.callee.name + '\': The first argument must not be preceded by any whitespace or comments (only \'(\').'
-			});
-
-			(lastCallArg.type !== 'NameValueAssignment' && charAfterLastCallArg !== ')') && context.report ({
-				node: callArgs [0],
-				location: {
-					column: sourceCode.getEndingColumn (lastCallArg) + 1
-				},
-				message: '\'' + node.callee.name + '\': The last argument must not be succeded by any whitespace or comments (only \')\').'
-			});
-		});
-
-
 		context.on ('NameValueAssignment', function (emitted) {
 			if (emitted.exit) {
 				return;
@@ -208,18 +136,6 @@ module.exports = {
 		context.on ('ObjectExpression', function (emitted) {
 			var node = emitted.node, properties = node.properties;
 
-			function inspectObjectPropForWhitespace (prop) {
-				var charAfterProp = sourceCode.getNextChar (prop);
-
-				(charAfterProp !== ',') && context.report ({
-					node: prop,
-					location: {
-						column: sourceCode.getEndingColumn (prop) + 1
-					},
-					message: 'There should be no whitespace or comments between object property and the comma following it.'
-				});
-			}
-
 			if (emitted.exit) {
 				return;
 			}
@@ -235,9 +151,6 @@ module.exports = {
 
 				return;
 			}
-
-			//CHECK FOR COMMA WHITESPACE
-			properties.slice (0, -1).forEach (inspectObjectPropForWhitespace);
 
 			var lastProperty = lastProperty = properties.slice (-1) [0];
 
@@ -265,331 +178,5 @@ module.exports = {
 				message: 'There should be no whitespace or comments between the last property and closing brace \'}\'.'
 			});
 		});
-
-
-		context.on ('AssignmentExpression', function (emitted) {
-			/**
-			 * node.operator is refined here by adding backslash before all the 'special' characters.
-			 * 'special' chars are thos chars that are part of solidity assignment operators and, if used without backslash in JS RegExp,
-			 * behave as wildcard characters. So to make sure they're treated as simple strings, we add '\' before them.
-			 * As of today, these chars include: * / + | ^
-			 */
-			var node = emitted.node,
-				op = node.operator.replace (/([\+\*\/\|\^])/g, '\\$1'), opLength = node.operator.length;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			// If expression is 'abc *= def;', then charsAfterLeftNode will contain ' *= d'.
-			var charsAfterLeftNode = sourceCode.getNextChars (node.left, 3 + opLength),
-				validationRegexp = new RegExp ('^ ' + op + ' [^\\s]$');
-
-			(!validationRegexp.test (charsAfterLeftNode)) && context.report ({
-				node: node.left,
-				message: 'Assignment operator must have exactly single space on both sides of it.'
-			});
-		});
-
-		//statement like `var x = 10` doesn't come under AssignmentExpression, so needs to be checked separately
-		context.on ('VariableDeclaration', function (emitted) {
-			var node = emitted.node, code = sourceCode.getText (node);
-
-			if (emitted.exit) {
-				return;
-			}
-
-			//if a particular character is '=', check its left and right for single space
-			for (var i = 2; i < code.length; i++) {
-				if (code [i] === '=') {
-					(!/^[^\/\s] $/.test (code.slice (i-2, i))) && context.report ({
-						node: node,
-						message: 'There should be only a single space between assignment operator \'=\' and its left side.' 
-					});
-
-					(!/^ [^\/\s]$/.test (code.slice (i+1, i+3))) && context.report ({
-						node: node,
-						message: 'There should be only a single space between assignment operator \'=\' and its right side.' 
-					});
-				}
-			}
-
-			//ensure there's no whitespace or comments before semicolon
-			(code [code.length - 1] === ';' && /(\s|\/)/.test (code [code.length - 2])) && context.report ({
-				node: node,
-				location: {
-					column: code.length - 2
-				},
-				message: 'There should be no whitespace or comments before the semicolon.'
-			});
-		});
-
-
-		//If we're dealing with abstract function declaration, we need to ensure no whitespce or comments before semicolon
-		context.on ('FunctionDeclaration', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			var code = sourceCode.getText (node);
-
-			(node.is_abstract && code [code.length - 1] === ';' && /(\s|\/)/.test (code [code.length - 2])) &&
-			context.report ({
-				node: node,
-				location: {
-					column: code.length - 2
-				},
-				message: 'There should be no whitespace or comments before the semicolon.'
-			});
-
-			//If parameters are specified, ensure appropriate spacing surrounding commas
-			var params = node.params;
-
-			if (params && params.length > 1) {
-				params.slice (0, -1).forEach (function (arg) {
-					sourceCode.getNextChar (arg) !== ',' && context.report ({
-						node: arg,
-						location: {
-							column: sourceCode.getEndingColumn (arg) + 1
-						},
-						message: node.name + ' ' + arg.id + '(): All arguments (except the last one) must be immediately followed by a comma.'
-					});
-				});
-			}
-		});
-
-
-		// The InformalParameter nodes (params) in modifier declaration should follow the same spacing rules as function declarations
-		context.on ('ModifierDeclaration', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			//If parameters are specified, ensure appropriate spacing surrounding commas
-			var params = node.params;
-
-			if (params && params.length > 1) {
-				params.slice (0, -1).forEach (function (arg) {
-					sourceCode.getNextChar (arg) !== ',' && context.report ({
-						node: arg,
-						location: {
-							column: sourceCode.getEndingColumn (arg) + 1
-						},
-						message: node.name + ' ' + arg.id + '(): All arguments (except the last one) must be immediately followed by a comma.'
-					});
-				});
-			}
-		});
-
-
-		context.on ('IfStatement', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			/**
-			 * Ensure a single space between 'if' token and the opening parenthesis 'if (...)'
-			 */
-			var ifTokenLength = 'if'.length,
-				nodeCode = sourceCode.getText (node).slice (ifTokenLength, ifTokenLength + 2);
-
-			(nodeCode !== ' (') && context.report ({
-				node: node,
-				location: {
-					column: sourceCode.getColumn (node) + ifTokenLength
-				},
-				message: 'There should be exactly a single space between the \'if\' token and the parenthetic block representing the conditional.'
-			});
-		});
-
-
-		context.on ('WhileStatement', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			/**
-			 * Ensure a single space between 'while' token and the opening parenthesis 'while (...)'
-			 */
-			var whileTokenLength = 'while'.length,
-				nodeCode = sourceCode.getText (node).slice (whileTokenLength, whileTokenLength + 2);
-
-			(nodeCode !== ' (') && context.report ({
-				node: node,
-				location: {
-					column: sourceCode.getColumn (node) + whileTokenLength
-				},
-				message: 'There should be exactly a single space between the \'while\' token and the parenthetic block representing the conditional.'
-			});
-		});
-
-
-		context.on ('ForStatement', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			/**
-			 * Ensure a single space between 'for' token and the opening parenthesis 'for (...)'
-			 */
-			var forTokenLength = 'for'.length,
-				nodeCode = sourceCode.getText (node).slice (forTokenLength, forTokenLength + 2);
-
-			(nodeCode !== ' (') && context.report ({
-				node: node,
-				location: {
-					column: sourceCode.getColumn (node) + forTokenLength
-				},
-				message: 'There should be exactly a single space between the \'for\' token and the parenthetic block representing the conditional.'
-			});
-		});
-
-
-		/************************************************************************************************
-		 * From here on, we're explicitly looking for semicolon whitespace errors
-		 ************************************************************************************************/
-		context.on ('ExpressionStatement', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			var code = sourceCode.getText (node);
-
-			//ensure there's no whitespace or comments before semicolon
-			(code [code.length - 1] === ';' && /(\s|\/)/.test (code [code.length - 2])) && context.report ({
-				node: node,
-				location: {
-					column: code.length - 2
-				},
-				message: 'There should be no whitespace or comments before the semicolon.'
-			});
-		});
-
-
-		context.on ('UsingStatement', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			var code = sourceCode.getText (node);
-
-			//ensure there's no whitespace or comments before semicolon
-			(code [code.length - 1] === ';' && /(\s|\/)/.test (code [code.length - 2])) && context.report ({
-				node: node,
-				location: {
-					column: code.length - 2
-				},
-				message: 'There should be no whitespace or comments before the semicolon.'
-			});
-		});
-
-
-		context.on ('ImportStatement', function (emitted) {
-			var node = emitted.node;
-
-			if (emitted.exit) {
-				return;
-			}
-
-			var code = sourceCode.getText (node);
-
-			//ensure there's no whitespace or comments before semicolon
-			(code [code.length - 1] === ';' && /(\s|\/)/.test (code [code.length - 2])) && context.report ({
-				node: node,
-				location: {
-					column: code.length - 2
-				},
-				message: 'There should be no whitespace or comments before the semicolon.'
-			});
-		});
-
-
-		/************************************************************************************************
-		 * From here on, we're explicitly looking for comma whitespace errors
-		 ************************************************************************************************/
-		context.on ('ArrayExpression', function (emitted) {
-			var node = emitted.node;
-
-			function inspectArrayElementForCommaWhitespace (elem) {
-				var charAfterElem = sourceCode.getNextChar (elem);
-
-				(charAfterElem !== ',') && context.report ({
-					node: elem,
-					location: {
-						column: sourceCode.getEndingColumn (elem) + 1
-					},
-					message: 'There should be no whitespace or comments between Array element and the comma following it.'
-				});
-			}
-
-			if (emitted.exit) {
-				return;
-			}
-
-			//need to test comma from starting to second last elem, since there's no comma after last element
-			node.elements.slice (0, -1).forEach (inspectArrayElementForCommaWhitespace);
-		});
-
-
-		context.on ('SequenceExpression', function (emitted) {
-			var node = emitted.node, expressions = node.expressions || [];
-
-			function inspectExprForWhitespace (expr) {
-				var charAfterExpr = sourceCode.getNextChar (expr);
-
-				(charAfterExpr !== ',') && context.report ({
-					node: expr,
-					location: {
-						column: sourceCode.getEndingColumn (expr) + 1
-					},
-					message: 'There should be no whitespace or comments between Tuple\'s element and the comma following it.'
-				});
-			}
-
-			if (emitted.exit) {
-				return;
-			}
-
-			expressions.slice (0, -1).forEach (inspectExprForWhitespace);
-		});
-
-
-		context.on ('VariableDeclarationTuple', function (emitted) {
-			var node = emitted.node, declarations = node.declarations;
-
-			function inspectVariableDeclaratorForWhitespace (vd) {
-				var charAfterExpr = sourceCode.getNextChar (vd);
-
-				(charAfterExpr !== ',') && context.report ({
-					node: vd,
-					location: {
-						column: sourceCode.getEndingColumn (vd) + 1
-					},
-					message: '\'' + vd.id.name + '\': identifier should be immediately followed by a comma without any whitespace in between.'
-				});
-			}
-
-			if (emitted.exit) {
-				return;
-			}
-
-			declarations.slice (0, -1).forEach (inspectVariableDeclaratorForWhitespace);
-		});
-
 	}
-
 };
